+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## 1.0.64 - 2026-07-10
+
 - fix(tests): ensure consistent formatting in import handler test. (`e38c9a3`)
 - fix(tests): update import handler test for consistent formatting. (`89509d1`)
 - feat(imports): add file import handler registration functionality. (`8d85a8f`)